Added cumulativeDeaths to Utils and changed graph colour to blue
