+++ frontend/src/Components/Results/Results.tsx
@@ -2,7 +2,7 @@ import { H3, H6, Divider, Spinner } from "@blueprintjs/core";
 import { useEffect, useState } from "react";
 import { GetResult } from "../../Helpers/API";
 import { Result } from "../../Helpers/Result";
-import { Average } from "../../Helpers/Utils";
+import { Average, addCumulativeDeaths } from "../../Helpers/Utils";
 import ConfigInfo from "./ConfigInfo";
 import ReportCard from "./ReportCard";
 import LineChart from "./Graphs/LineChart";
@@ -47,52 +47,7 @@ interface ResultDisplayProps {
 }
 function ResultDisplay(props: ResultDisplayProps) {
   const { result } = props;
-
-  // CUMULATIVE DEATHS PER AGENT CODE
-  // Finding which agent types are duplicated and their respective indices
-  // Adding the cumulative deaths, and removing duplicated agentTypes
-  var agentTypes = result.deaths.map((d) => d.agentType);
-  var cumulativeDeaths = result.deaths.map((d) => d.cumulativeDeaths);
-  var myMap = new Map<string, number[]>();
-
-  for (var i = 0; i < agentTypes.length; i++) {
-    var element = agentTypes[i].toString();
-    var test = myMap.get(element);
-    if (typeof test != "undefined") {
-      test.push(i);
-      myMap.set(element, test);
-    } else {
-      myMap.set(element, [i]);
-    }
-  }
-
-  var findDuplicates = Array.from(myMap.entries()).filter(([key, val]) => {
-    if (val.length > 1) {
-      return [key, val];
-    }
-  });
-
-  if (findDuplicates) {
-    for (var j = 0; j < findDuplicates.length; j++) {
-      let [n, idx] = findDuplicates[j];
-      var tempValue = 0;
-      for (var i = 0; i < idx.length; i++) {
-        tempValue += cumulativeDeaths[idx[i]];
-      }
-      cumulativeDeaths[idx[0]] = tempValue;
-
-      for (var z = 1; z < idx.length; z++) {
-        delete cumulativeDeaths[idx[z]];
-        delete agentTypes[idx[z]];
-      }
-    }
-    cumulativeDeaths = cumulativeDeaths.filter(function (e) {
-      return e;
-    });
-    agentTypes = agentTypes.filter(function (e) {
-      return e;
-    });
-  }
+  let { agentTypes, cumulativeDeaths } = addCumulativeDeaths(result.deaths);
 
   return (
     <>
@@ -115,7 +70,7 @@ function ResultDisplay(props: ResultDisplayProps) {
       </div>
       <div className="row">
         <div className="col-lg-6">
-          <BarChart yAxis={cumulativeDeaths} xAxis={agentTypes} graphTitle="Cumulative Deaths per Agent type" />
+          <BarChart yAxis={cumulativeDeaths} xAxis={agentTypes} graphTitle="Deaths per Agent type" />
         </div>
         <div className="col-lg-6">
           <LineChart
